+++ src/spmkit/gui/viewmodels/__init__.py
@@ -1,9 +1,12 @@
 """ViewModels (MVVM): estado observable entre el ``core`` puro y los paneles Qt."""
 
 from spmkit.gui.viewmodels.batch_vm import BatchViewModel
+from spmkit.gui.viewmodels.figure_vm import FigureViewModel
 from spmkit.gui.viewmodels.force_vm import DEFAULT_RECIPE, ForceViewModel
 from spmkit.gui.viewmodels.image_vm import ImageViewModel
 from spmkit.gui.viewmodels.map_vm import PROPERTIES, MapViewModel
+from spmkit.gui.viewmodels.simulator_vm import SimulatorViewModel
+from spmkit.gui.viewmodels.view3d_vm import View3DViewModel
 
 __all__ = [
     "ForceViewModel",
@@ -12,4 +15,7 @@ __all__ = [
     "PROPERTIES",
     "BatchViewModel",
     "ImageViewModel",
+    "FigureViewModel",
+    "SimulatorViewModel",
+    "View3DViewModel",
 ]